Cloning repos spawned into a goroutine and done lazily in the background
Launch cloneRepos into a goroutine that ranges over the new updateRepoCh
channel. An error on cloning or fetching updates on a repo is no longer
fatal, but will print an error to stdout for now. config.Repo's
ClonedRepo attribute is now a pointer to repository.Repository so we can
share the cloned repo without having to reassign the entire config.Repo
to the global map.